requirements.txt
librosa>=0.10.0
numpy>=1.24.0
Pillow>=10.0.0
moviepy>=1.0.3
customtkinter>=5.2.0
pyinstaller>=6.0.0
